backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
sqlmodel==0.0.22
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
websockets==12.0
# Discord
discord.py==2.4.0
# OpenAI クライアント（抽象化の上で任意プロバイダに差し替え可）
openai==1.42.0
# Twitter スクレイピング
twitter-api-client==0.10.21
# Cookie自動取得（オプション）
playwright==1.49.0
selenium==4.27.1
webdriver-manager==4.0.2
futu-api
# Alpaca Markets
alpaca-py>=0.31.0


